+++ StravaKit/Route.swift
@@ -1,7 +1,7 @@
 
 public struct Route: Codable, Equatable {
 
-    public let id: Int
+    public let id: Int64
     public let name: String
     public let description: String
     public let map: Map
