+++ StravaKit/UploadStatus.swift
@@ -1,11 +1,11 @@
 
 public struct UploadStatus: Codable, Equatable {
 
-    public let id: Int
+    public let id: Int64
     public let externalId: String
     public let error: String?
     public let status: String
-    public let activityId: Int?
+    public let activityId: Int64?
 
     enum CodingKeys : String, CodingKey {
         case id
